Complete Java Maps Project: Employee Deduplication and Word Counter Implementation
Employee class was created: Employee objects with id, firstname and lastname fields and a constructor with parameters were defined.
findDuplicates method: A solution was provided to detect duplicate Employee objects in LinkedList, add them to another list and return this list.
findUniques method: Takes only one of the repeating elements in the LinkedList and all non-repeating elements and adds them to a Map structure.
removeDuplicates method: Removed all duplicates in LinkedList and returned only the unique ones.
WordCounter class: Contains a calculateWord method that finds the number of times each word is repeated in the given text. By returning the results in Map<String, Integer> format, the word counts can be easily analysed.
Solutions were successfully completed using LinkedList and HashMap in accordance with the project requirements.